+++ typescript-api/src/store/Products.ts
@@ -89,6 +89,7 @@ export class Products extends VuexModule {
     } catch (error) {
       this.context.commit("SET_ERROR_STATUS", true);
     } finally {
+      console.log("created");
       this.context.commit("SET_LOADING_STATUS", false);
     }
   }
